Tried debugging polygon function
The returnPolygon function now works to fetch and construct a string of the lat long bounds of the polygon. The structure of the overpass_query in the polygonInfo function should be identical to the one in shardul's python code, but somehow I got this error ""runtime error: Query timed out in "print" at line 1 after 183 seconds."
If that's my network problem, the code should work and it should download the result from overpass api into a text file.